INTERMEDIATE CODE: Kuka iiwa driver, major steps towards fully configurable system.

+++ src/java/grl/src/grl/flatbuffer/KUKAiiwaInterface.java
@@ -4,10 +4,10 @@ package grl.flatbuffer;
 
 public final class KUKAiiwaInterface {
   private KUKAiiwaInterface() { }
-  public static final short Disabled = 0;
-  public static final short SmartServo = 1;
-  public static final short DirectServo = 2;
-  public static final short FRI = 3;
+  public static final byte Disabled = 0;
+  public static final byte SmartServo = 1;
+  public static final byte DirectServo = 2;
+  public static final byte FRI = 3;
 
   private static final String[] names = { "Disabled", "SmartServo", "DirectServo", "FRI", };
